feat(codemod): getTableData => useTableData for util
