+++ packages/codemod/transforms/__tests__/obutil-to-oceanbase-util.test.ts
@@ -1,7 +1,7 @@
 import { defineTest } from 'jscodeshift/src/testUtils';
 
 const testUnit = 'obutil-to-oceanbase-util';
-const tests = ['obutil'];
+const tests = ['obutil', 'getTableData'];
 
 describe(testUnit, () => {
   tests.forEach(test =>
